[FEAT] US 02 - Endpoint 01 02
Implementação de endpoints:
01 */fresh-products*
+ Listagem de produtos completa
+ exceção customizada 404 de listagens retorno nulo 'empty list'

02 */fresh-products/list?categoria=[categoriaId]*
Listagem de produtos por filtro de Categoria
[{'FS': 'FRESCO','RR': 'REFRIGERADO','FF': 'CONGELADO'}]

Modelagem de banco de dados:
+ Produto <-> Categoria: 2 Char
+ Produto <-> Vendedor: chave composta e atributo preço em relação
+ Produto e preço são atualizados a partir de endpoint lote via ProdutoDTO

teste:
+ ProdutoService

+++ src/main/java/br/com/meliw4/projetointegrador/dto/response/ProdutoResponseDTO.java
@@ -12,7 +12,7 @@ import lombok.Data;
 public class ProdutoResponseDTO {
 	private Long id;
 	private final LocalDate dataVencimento;
-	private final Integer volume;
+	private final Double volume;
 	private final ProdutoCategoria produtoCategoria;
 
 	public static ProdutoResponseDTO toDTO(Produto produto) {
